Deffered Material 작업 완료, Animator쪽 작업 필요

+++ Engine/GPU/StructBuffer.h
@@ -42,9 +42,9 @@ namespace ehw
 
         //desc를 바꾸고자 할 경우 그냥 구조화 버퍼를 지웠다가 다시 생성할것.
         template <typename T>
-        eResult Init(const Desc& _tDesc, uint _capacity = 0, const void* _pInitialData = nullptr, size_t _ElemCount = 0);
+        eResult Init(const Desc& _tDesc, uint _capacity = 0, const void* _pInitialData = nullptr, UINT _ElemCount = 0);
 
-        eResult Resize(size_t _ElementCapacity, const void* _pInitialData = nullptr, size_t _dataCount = 0);
+        eResult Resize(size_t _ElementCapacity, const void* _pInitialData = nullptr, UINT _dataCount = 0);
 
         //Setter Getter Adder
         void SetPipelineTarget(eShaderStageFlag_ _StageFlag) { m_desc.TargetStageSRV = _StageFlag; }
@@ -58,7 +58,7 @@ namespace ehw
         uint GetStride() const { return m_elementStride; }
 
         //데이터를 버퍼로 전송
-        void SetData(const void* _pData, size_t _uCount);
+        void SetData(const void* _pData, UINT _uCount);
 
         //데이터를 받아옴
         void GetData(void* _pDest, size_t _uDestByteCapacity);
@@ -82,9 +82,9 @@ namespace ehw
     private:
         Desc                        m_desc;
 
-        uint                        m_elementStride;   //구조체 하나 당 바이트 갯수
-        uint                        m_elementCount;    //현재 등록한 구조체의 갯수
-        uint                        m_capacity; //현재 확보되어있는 구조체의 갯수
+        UINT                        m_elementStride;   //구조체 하나 당 바이트 갯수
+        UINT                        m_elementCount;    //현재 등록한 구조체의 갯수
+        UINT                        m_capacity; //현재 확보되어있는 구조체의 갯수
 
         ComPtr<ID3D11ShaderResourceView> m_SRV;
 
@@ -98,12 +98,12 @@ namespace ehw
 	};
 
     template<typename T>
-    inline eResult StructBuffer::Init(const Desc& _tDesc, uint _capacity, const void* _pInitialData, size_t _elemCount)
+    inline eResult StructBuffer::Init(const Desc& _tDesc, uint _capacity, const void* _pInitialData, UINT _elemCount)
     {
         SetDesc(_tDesc);
 
         static_assert(sizeof(T) % 16 == 0, "The byte size of the structured buffer must be a multiple of 16.");
-        m_elementStride = (uint)sizeof(T);
+        m_elementStride = (UINT)sizeof(T);
 
         return Resize(_capacity, _pInitialData, _elemCount);
     }
